requirements.txt
requests_oauthlib==1.3.0
pyyaml
python-dateutil==2.8.1
requests